more elegant solution
